handle_error

+++ pojo-ops/src/main/java/com/stanzaliving/core/opscalculator/dto/HouseKeepingInfoDto.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class HouseKeepingInfoDto {
     private int employeesCount;
     private int ratePerMonth;
-    private int leaveFactor;
+    private double leaveFactor;
 }
